TPASPKT-391 Separate user and project DB entries
Store user and project information in two separate MongoDB collections
instead of in one. This comes with multiple changes:

* User is no longer required to have any projects assigned by default.
* Access to project(s) can be granted and revoked after creating the
  user.
* User can have access to multiple projects.
* Project ID has to be provided as part of the URL when accessing
project-related API endpoints.

+++ upload_rest_api/api/v1/files.py
@@ -13,17 +13,18 @@ import upload_rest_api.database as db
 import upload_rest_api.gen_metadata as md
 import upload_rest_api.upload as up
 import upload_rest_api.utils as utils
+from upload_rest_api.authentication import current_user
 from upload_rest_api.api.v1.tasks import TASK_STATUS_API_V1
 from upload_rest_api.jobs.utils import FILES_QUEUE, enqueue_background_job
 
 FILES_API_V1 = Blueprint("files_v1", __name__, url_prefix="/v1/files")
 
 
-def _get_dir_tree(user, fpath):
+def _get_dir_tree(project_id, fpath):
     """Return with dir tree from fpath as a dict."""
     file_dict = {}
     for dirpath, _, files in os.walk(fpath):
-        path = utils.get_return_path(user, dirpath)
+        path = utils.get_return_path(project_id, dirpath)
         file_dict[path] = files
 
     if "." in file_dict:
@@ -32,20 +33,24 @@ def _get_dir_tree(user, fpath):
     return file_dict
 
 
-@FILES_API_V1.route("/<path:fpath>", methods=["POST"])
-def upload_file(fpath):
+@FILES_API_V1.route("/<string:project_id>/<path:fpath>", methods=["POST"])
+def upload_file(project_id, fpath):
     """Save the uploaded file at <UPLOAD_PATH>/project/fpath.
 
     :returns: HTTP Response
     """
+    if not current_user.is_allowed_to_access_project(project_id):
+        abort(403, "No permission to access this project")
+
     database = db.Database()
-    user = database.user(request.authorization.username)
     upload_path = safe_join("", fpath)
 
-    up.validate_upload(user, request.content_length, request.content_type)
+    up.validate_upload(
+        project_id, request.content_length, request.content_type
+    )
 
     md5 = up.save_file(database,
-                       user,
+                       project_id,
                        request.stream,
                        request.args.get('md5', None),
                        upload_path)
@@ -59,21 +64,36 @@ def upload_file(fpath):
     )
 
 
-@FILES_API_V1.route("/", defaults={'fpath': ""}, methods=["GET"])
-@FILES_API_V1.route("/<path:fpath>", methods=["GET"])
-def get_path(fpath):
+@FILES_API_V1.route(
+    "/<string:project_id>/", defaults={'fpath': ""}, methods=["GET"],
+    strict_slashes=False
+)
+@FILES_API_V1.route(
+    "/<string:project_id>/<path:fpath>", methods=["GET"]
+)
+def get_path(project_id, fpath):
     """Get filepath, name and checksum.
 
     :returns: HTTP Response
     """
-    username = request.authorization.username
+    if not current_user.is_allowed_to_access_project(project_id):
+        abort(403, "No permission to access this project")
+
     database = db.Database()
-    user = database.user(username)
 
-    upload_path = utils.get_upload_path(user, fpath)
-    return_path = utils.get_return_path(user, upload_path)
+    upload_path = utils.get_upload_path(project_id, fpath)
+    return_path = utils.get_return_path(project_id, upload_path)
 
-    if os.path.isfile(upload_path):
+    if request.args.get("all", None) == "true" and fpath.strip("/") == "":
+        # Retrieve entire directory listing if 'all' URL parameter is set
+        # and fpath is set to '' or '/'
+        fpath = db.Projects.get_project_directory(project_id)
+
+        if not os.path.exists(fpath):
+            abort(404, "No files found")
+
+        response = jsonify(_get_dir_tree(project_id, fpath))
+    elif os.path.isfile(upload_path):
         response = {
             "file_path": return_path,
             "identifier": database.files.get_identifier(upload_path),
@@ -91,8 +111,9 @@ def get_path(fpath):
             verify=current_app.config.get("METAX_SSL_VERIFICATION")
         )
         try:
-            identifier = metax.get_project_directory(user.get_project(),
-                                                     return_path)['identifier']
+            identifier = metax.get_project_directory(
+                project_id, return_path
+            )['identifier']
         except metax_access.DirectoryNotAvailableError:
             identifier = None
 
@@ -112,28 +133,32 @@ def get_path(fpath):
     return response
 
 
-@FILES_API_V1.route("/", defaults={'fpath': ""}, methods=["DELETE"],
-                    strict_slashes=False)
-@FILES_API_V1.route("/<path:fpath>", methods=["DELETE"])
-def delete_path(fpath):
-    """Delete fpath under user's project.
+@FILES_API_V1.route(
+    "/<string:project_id>", defaults={'fpath': ""}, methods=["DELETE"],
+    strict_slashes=False
+)
+@FILES_API_V1.route("/<string:project_id>/<path:fpath>", methods=["DELETE"])
+def delete_path(project_id, fpath):
+    """Delete fpath under project.
 
     If fpath resolves to a directory, the whole directory is recursively
     removed.
 
     :returns: HTTP Response
     """
+    if not current_user.is_allowed_to_access_project(project_id):
+        abort(403, "No permission to access this project")
+
     root_upload_path = current_app.config.get("UPLOAD_PATH")
-    username = request.authorization.username
     database = db.Database()
-    user = database.user(username)
-    upload_path = utils.get_upload_path(user, fpath)
+    upload_path = utils.get_upload_path(project_id, fpath)
+    project_dir = database.projects.get_project_directory(project_id)
 
     if os.path.isfile(upload_path):
         # Remove metadata from Metax
         try:
             response = md.MetaxClient().delete_file_metadata(
-                user.get_project(), upload_path, root_upload_path
+                project_id, upload_path, root_upload_path
             )
         except md.MetaxClientError as exception:
             response = str(exception)
@@ -143,7 +168,7 @@ def delete_path(fpath):
         os.remove(upload_path)
 
     elif upload_path.exists() \
-            and upload_path.samefile(user.project_directory) \
+            and upload_path.samefile(project_dir) \
             and not any(upload_path.iterdir()):
         # Trying to delete empty project directory
         abort(404, "No files found")
@@ -153,16 +178,16 @@ def delete_path(fpath):
         task_id = enqueue_background_job(
             task_func="upload_rest_api.jobs.files.delete_files",
             queue_name=FILES_QUEUE,
-            username=username,
+            project_id=project_id,
             job_kwargs={
                 "fpath": upload_path,
-                "username": username
+                "project_id": project_id
             }
         )
 
         polling_url = utils.get_polling_url(TASK_STATUS_API_V1.name, task_id)
         response = jsonify({
-            "file_path": utils.get_return_path(user, upload_path),
+            "file_path": utils.get_return_path(project_id, upload_path),
             "message": "Deleting files and metadata",
             "polling_url": polling_url,
             "status": "pending"
@@ -176,31 +201,13 @@ def delete_path(fpath):
     else:
         abort(404, "File not found")
 
-    database.user(username).update_used_quota(root_upload_path)
+    database.projects.update_used_quota(project_id, root_upload_path)
 
     response = jsonify({
-        "file_path": utils.get_return_path(user, upload_path),
+        "file_path": utils.get_return_path(project_id, upload_path),
         "message": "deleted",
         "metax": response
     })
     response.status_code = 200
 
-    return response
-
-
-@FILES_API_V1.route("", methods=["GET"])
-def get_files():
-    """Get all files of the user.
-
-    :return: HTTP Response
-    """
-    username = request.authorization.username
-    user = db.Database().user(username)
-    fpath = user.project_directory
-
-    if not os.path.exists(fpath):
-        abort(404, "No files found")
-
-    response = jsonify(_get_dir_tree(user, fpath))
-    response.status_code = 200
     return response
